requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
bootstrap-flask==2.3.0
python-dotenv==1.0.1
openai==1.12.0
httpx==0.27.0 